Extract shared SIMD kernels into `simd_generic` module
Make the structure of the `gemm::kernels` entry point a bit easier to follow by
moving the `simd_*` functions into a submodule that is imported by the various
architecture-specific modules.
