+++ src/gemm/kernels/generic.rs
@@ -1,9 +1,11 @@
 use std::mem::MaybeUninit;
 use std::ops::Range;
 
+use rten_simd::vec_count;
 use rten_tensor::Matrix;
 
-use super::{simd_gemm, simd_gemv, vec_count, Kernel};
+use super::simd_generic::{simd_gemm, simd_gemv};
+use super::Kernel;
 use crate::gemm::packing::{pack_a_block, pack_b_block};
 
 /// This is the base kernel that does not use architecture-specific intrinsics
